move configs out to separate files
